Dev talha

+++ src/main/java/com/osudpotro/posmaster/offerhub/membership/MembershipCreateRequest.java
@@ -2,10 +2,12 @@ package com.osudpotro.posmaster.offerhub.membership;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class MembershipCreateRequest {
     private String name;
-    private Double discount;
+    private BigDecimal discount;
     private Boolean isPercentage;
     private Double maxDiscount;
 }
